refactor and improvements
