+++ integration/restValidation.ts
@@ -2,8 +2,8 @@ import {init as initBLS, SecretKey} from "@chainsafe/bls";
 import {LogLevel, WinstonLogger} from "@chainsafe/lodestar-utils";
 import {Validator} from "@chainsafe/lodestar-validator";
 import {config} from "@chainsafe/lodestar-config/lib/presets/mainnet";
-import {CgEth2ApiClient} from "../../src/renderer/services/eth2/client/eth2ApiClient";
-import {CGSlashingProtection} from "../../src/renderer/services/eth2/client/slashingProtection";
+import {CgEth2ApiClient} from "../src/renderer/services/eth2/client/eth2ApiClient";
+import {CGSlashingProtection} from "../src/renderer/services/eth2/client/slashingProtection";
 import sinon from "sinon";
 import EventSource from "eventsource";
 import {computeEpochAtSlot} from "@chainsafe/lodestar-beacon-state-transition";
