+++ package.json
@@ -57,7 +57,9 @@
     "script": "node -r ts-node/register",
     "generate:validators": "yarn run script scripts/generate_interop_validators.ts",
     "testnet:dev": "./testnet/lighthouse/reset_genesis_time.sh && docker-compose -f lighthouse-testnet.yml up",
-    "build-storybook": "build-storybook"
+    "build-storybook": "build-storybook",
+    "integration:lighthouse": "yarn run script integration/lighthouse.ts",
+    "integration:teku": "yarn run script integration/teku.ts.ts"
   },
   "jest": {
     "transform": {
