+++ src/renderer/services/eth2/client/eth2ApiClient/cgEth2BeaconApi.ts
@@ -8,6 +8,7 @@ import {CgEth2BeaconBlocksApi} from "./cgEth2BeaconBlocksApi";
 import {CgEth2BeaconStateApi} from "./cgEth2BeaconStateApi";
 import {CgEth2BeaconPoolApi} from "./cgEth2BeaconPoolApi";
 import logger from "electron-log";
+import {Dispatch} from "redux";
 
 export class CgEth2BeaconApi implements ICGEth2BeaconApi {
     public blocks: ICGETH2BeaconBlocksApi;
@@ -16,13 +17,13 @@ export class CgEth2BeaconApi implements ICGEth2BeaconApi {
 
     private readonly httpClient: HttpClient;
     private readonly config: IBeaconConfig;
-    public constructor(config: IBeaconConfig, httpClient: HttpClient, publicKey?: string) {
+    public constructor(config: IBeaconConfig, httpClient: HttpClient, publicKey?: string, dispatch?: Dispatch) {
         this.config = config;
         this.httpClient = httpClient;
 
         this.blocks = new CgEth2BeaconBlocksApi(config, httpClient);
         this.state = new CgEth2BeaconStateApi(config, httpClient);
-        this.pool = new CgEth2BeaconPoolApi(config, httpClient, publicKey);
+        this.pool = new CgEth2BeaconPoolApi(config, httpClient, publicKey, dispatch);
     }
 
     public getGenesis = async (): Promise<Genesis | null> => {
